Move secret_tests into config_tests
